requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
bcrypt==3.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.10.7
crypto==1.4.1
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
Faker==37.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
google-auth==2.32.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Naked==0.1.32
orjson==3.10.5
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
